requirements.txt
streamlit
pandas
plotly
google-cloud-bigquery
db-dtypes